updated .dat processing capabilities

+++ preProcessing.py
@@ -20,36 +20,36 @@ from datetime import datetime
 import time
 importlib.reload(parseFile)
 
-class read_ALL():
-    def __init__(self,SiteID,Year,Month,processes=1,reset=0,Test=0):
 
+class read_ALL():
+    def __init__(self,SiteID,Year,Month,processes=1,reset=0,Test=0,file_type='ghg',copy_From=None,metadata_template='None'):
+        self.file_type=file_type
         self.Year = str(Year)
         self.Month = "{:02d}".format(Month)
         self.SiteID = SiteID
+        self.copy_From = copy_From
 
         # Concatenate and read the ini files
         inis = ['configuration.ini','Metadata_Instructions.ini']
         ini_file = ['ini_files/'+ini for ini in inis]
         self.ini = configparser.ConfigParser()
         self.ini.read(ini_file)
+        self.Root = sub_path(self,self.ini['Paths']['root_data_dir'])
         
-        # self.MetadataTemplate = configparser.ConfigParser()
         # Create directories from template (see configuration.ini)
         self.ini['Paths']['dpath'] = sub_path(self,self.ini['Paths']['raw'])
         self.ini['Paths']['meta_dir'] = sub_path(self,self.ini['Paths']['metadata'])
         self.ini['Paths']['biomet_data'] = sub_path(self,self.ini['Paths']['biomet']+self.ini['filenames']['biomet'])
-        
-        self.ini['templates']['UpdateMetadata'] = self.ini['templates']['UpdateMetadata'].replace('METADATA',self.ini['Paths']['meta_dir'])
+        self.ini['filenames']['metadata_to_overwrite'] = sub_path(self,self.Root+self.ini['filenames']['metadata_to_overwrite'])
+        self.ini['templates']['metadata_template']=metadata_template
 
         if reset == 1:
-            try:
+            proceed = input(f"Warning: You are about to complete a hard reset, deleting all contents of: {self.ini['Paths']['meta_dir']}\nDo you wish to proceed? Y/N")
+            if proceed.lower() == 'y' and os.path.isdir(self.ini['Paths']['meta_dir']):
+                print(f"Deleting all contents of: {self.ini['Paths']['meta_dir']}")
                 shutil.rmtree(self.ini['Paths']['meta_dir'])
-            except:
-                pass
-            try:
-                os.mkdir(self.ini['Paths']['meta_dir'])
-            except:
-                pass
+            else:
+                print(f"Keeping contents of: {self.ini['Paths']['meta_dir']}")
             
         self.Parser = parseFile.Parse(self.ini)
 
@@ -57,36 +57,43 @@ class read_ALL():
             Path(self.ini['Paths']['meta_dir']).mkdir(parents=True, exist_ok=True)
         self.Logs = []
         
-        # Check for new .ghg files - or overwrite if rest flag is set to true
+        # Check for new .files - or overwrite if reset flag is set to true
         self.find_files(reset)
         if hasattr(self, 'files'):
             if self.files.loc[self.files.index.month==int(self.Month)].size>0:
                 self.Read(processes,Test,reset)
 
     def find_files (self,reset=0):
-        # Find the name of every FULL .ghg or .dat file in the raw data folder located at the end of directory tree
+        # Find the name of every FULL file in the raw data folder located at the end of directory tree
         # Exclude any files that fall off half hourly intervals ie. maintenance
         # Resample to 30 min intervals - missing filenames will be null
         # Save the list of files
 
-        # Get every .ghg or .dat file that is at the end of a directory tree
+        # Get every file that is at the end of a directory tree
         all_files = []
-        dates = []
+        TIMESTAMP = []
+        name_pattern = []
+        if self.copy_From is not None:
+            if os.path.isdir(self.ini['Paths']['dpath']) == False:
+                os.mkdir(self.ini['Paths']['dpath'])
+            self.setup()
         if os.path.isdir(self.ini['Paths']['dpath']):
             for file in os.listdir(self.ini['Paths']['dpath']):
-                name, tag = file.rsplit('.',1)
-                if (tag == 'ghg' or tag == 'dat'):
+                if file.endswith(self.file_type):
+                    name = file.rsplit('.',1)[0]
                     all_files.append(file)
-                    dates.append(pd.to_datetime(re.search(self.ini['FileName_DateFormat'][tag], name)[0].replace('_','-')))
+                    timestamp_info = re.search(self.ini[self.file_type]['search'], name).group(0)
+                    TIMESTAMP.append(datetime.strptime(timestamp_info,self.ini[self.file_type]['format']))
+                    name_pattern.append(name.replace(timestamp_info,self.ini[self.file_type]['ep_date_pattern']+'.'+self.file_type))
                     
             # Create dataframe of all GHG files
-            df = pd.DataFrame(data={'filename':all_files,'TIMESTAMP':dates})
+            df = pd.DataFrame(data={'filename':all_files,'TIMESTAMP':TIMESTAMP,'name_pattern':name_pattern})
             df = df.set_index('TIMESTAMP')
             df['Flag']='-'
             df['MetaDataFile'] = '-'
             df['Update']='-'
             df['setup_ID']=np.nan
-            
+
             # Flag timestamps with incomplete records
             df.loc[(((df.index.minute!=30)&(df.index.minute!=0))|(df.index.second!=0)),'Flag'] = 'Incomplete Record'
             for i,row in df.loc[((df['Flag']=='Incomplete Record')&(~df['filename'].str.contains('_incomplete')))].iterrows():
@@ -106,9 +113,18 @@ class read_ALL():
                 self.files = pd.concat([self.files,df.loc[df.index.isin(self.files.index)==False]])
                 # Sort so that oldest files get processed first
                 self.files = self.files.sort_index()#ascending=False)
-        
         else:
             print(f"Not a valid directory: {self.ini['Paths']['dpath']}")
+
+    def setup(self,path_in,path_out,date_format):
+        for _, _, files in os.walk(path_in):
+            for file in files:
+                if file.endswith(self.file_type):
+                    srch = re.search(self.ini[self.file_type]['search'], file).group(0)
+                    if srch is not None:
+                        TIMESTAMP =  datetime.strptime(srch.group(0),date_format)
+                        if TIMESTAMP.year == self.Year and str(TIMESTAMP.month).zfill(2) == self.Month:
+                            shutil.copy(f"{self.copy_From}/{file}",f"{self.ini['Paths']['dpath']}/{file}")
 
 
     def makeEmpty(self,type='object',ixName='TIMESTAMP'):
@@ -135,9 +151,9 @@ class read_ALL():
         self.Log = self.makeEmpty()
 
         if Test > 0:
-            self.to_Process = self.files.loc[(self.files['setup_ID'].isna()&((self.files['filename'].str.endswith('.ghg')|(self.files['filename'].str.endswith('.dat')))))][:Test]
+            self.to_Process = self.files.loc[(self.files['setup_ID'].isna()&(self.files['filename'].str.endswith(self.file_type)))][:Test]
         else:
-            self.to_Process = self.files.loc[(self.files['setup_ID'].isna()&((self.files['filename'].str.endswith('.ghg')|(self.files['filename'].str.endswith('.dat')))))]
+            self.to_Process = self.files.loc[(self.files['setup_ID'].isna()&(self.files['filename'].str.endswith(self.file_type)))]
 
         NameList = self.to_Process['filename'].values
         TimeList = self.to_Process.index
@@ -145,8 +161,8 @@ class read_ALL():
             # Run once to get the first Metadata file and use it as a template
             self.Parser.process_file([NameList[0],TimeList[0]],Template_File=False)
         if len(NameList)>0:
-            pb = progressbar(len(NameList),'Preprocessing Files')
             if (__name__ == 'preProcessing' or __name__ == '__main__') and processes>1:
+                pb = progressbar(len(NameList),'Preprocessing Files')
                 with Pool(processes=processes) as pool:
 
                     # Break year into one-day chunks
@@ -161,10 +177,9 @@ class read_ALL():
             # Sequential processing is helpful for trouble shooting but will run much slower
             else:
                 for fn,ts in zip(NameList,TimeList):
-                    # print(fn,ts)
                     out = self.Parser.process_file([fn,ts],Testing=True)
                     self.appendRecs(out)
-                    pb.step()
+                    # pb.step()
                     self.out = out
 
             self.prepareOutputs()        
@@ -176,6 +191,8 @@ class read_ALL():
         dyn = self.ini['Monitor']['dynamic_metadata'].split(',')
         self.dynamic_metadata = pd.concat([self.dynamic_metadata,df[dyn]])
         stup = self.ini['Monitor']['site_setup'].split(',')
+        stup = [x for x in stup if x in df.columns.values]
+        self.ini['Monitor']['site_setup'] = ','.join(stup)
         self.site_setup = pd.concat([self.site_setup,df[stup]])
 
         self.dataRecords = pd.concat([self.dataRecords,df.loc[:,df.columns.isin(dyn+stup)==False]])
@@ -200,6 +217,7 @@ class read_ALL():
 
     def GroupCommon(self):
         self.site_setup['temp_ID'] = self.site_setup.reset_index().index
+        self.site_setup = self.site_setup.fillna(0)
         self.site_setup['temp_ID'] = self.site_setup.groupby(self.ini['Monitor']['site_setup'].split(','))[['temp_ID']].transform(lambda x: x.min())
         self.site_setup['setup_ID']=np.nan
         for i,val in enumerate(self.site_setup['temp_ID'].unique()):
@@ -242,7 +260,7 @@ if __name__ == '__main__':
     "--SiteID", 
     nargs='+', # 1 or more values expected => creates a list
     type=str,
-    default=['BB','BB2','BBS','RBM','DSM','HOGG','YOUNG'],
+    default=[],
     )
 
     CLI.add_argument(
